# Conflicts:
#	src/main/java/com/niitcoder/coursegrade/service/impl/CourseHomeworkServiceImpl.java

+++ src/main/java/com/niitcoder/coursegrade/service/impl/CourseGroupServiceImpl.java
@@ -1,28 +1,21 @@
 package com.niitcoder.coursegrade.service.impl;
 
-import com.niitcoder.coursegrade.domain.CourseInfo;
+import com.niitcoder.coursegrade.domain.*;
+import com.niitcoder.coursegrade.repository.*;
+import com.niitcoder.coursegrade.security.SecurityUtils;
 import com.niitcoder.coursegrade.service.CourseGroupService;
-import com.niitcoder.coursegrade.domain.CourseGroup;
-import com.niitcoder.coursegrade.repository.CourseGroupRepository;
-import com.alibaba.fastjson.util.TypeUtils;
-import com.niitcoder.coursegrade.service.dto.CourseGroupDTO;
+import com.niitcoder.coursegrade.service.CourseInfoService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-
-import java.time.ZoneId;
 import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 /**
@@ -35,13 +28,31 @@ public class CourseGroupServiceImpl implements CourseGroupService {
     private final Logger log = LoggerFactory.getLogger(CourseGroupServiceImpl.class);
 
     private final CourseGroupRepository courseGroupRepository;
-    private final JdbcTemplate jdbcTemplate;
+    private final StudentCourseGroupRepository studentCourseGroupRepository;
+    private final CourseInfoRepository courseInfoRepository;
+    private final CourseInfoService courseInfoService;
 
-    public CourseGroupServiceImpl(CourseGroupRepository courseGroupRepository, JdbcTemplate jdbcTemplate) {
+    public CourseGroupServiceImpl(CourseGroupRepository courseGroupRepository, StudentCourseGroupRepository studentCourseGroupRepository,CourseInfoRepository courseInfoRepository, CourseInfoService courseInfoService) {
         this.courseGroupRepository = courseGroupRepository;
-        this.jdbcTemplate = jdbcTemplate;
+        this.studentCourseGroupRepository = studentCourseGroupRepository;
+        this.courseInfoRepository = courseInfoRepository;
+        this.courseInfoService = courseInfoService;
     }
 
+    /**
+     * 检查班级是否存在,存在，则返回查到的班级，不存在，则抛出异常
+     * @param id
+     * @return
+     * @throws Exception
+     */
+    public CourseGroup isExistCourseInfo(Long id) throws Exception {
+        //检查班级是否存在
+        Optional<CourseGroup> courseGroup=courseGroupRepository.findById(id);
+        if(!courseGroup.isPresent()){
+            throw new Exception("班级不存在.");
+        }
+        return courseGroup.get();
+    }
     /**
      * Save a courseGroup.
      *
@@ -49,24 +60,61 @@ public class CourseGroupServiceImpl implements CourseGroupService {
      * @return the persisted entity.
      */
     @Override
-    public CourseGroup save(CourseGroup courseGroup) {
+    public CourseGroup save(CourseGroup courseGroup,Long courseInfoId) throws Exception {
         log.debug("Request to save CourseGroup : {}", courseGroup);
+        //检查课程是否由该用户创建，若不是，则无权限创建和更改此课程下的班级
+        if(!courseInfoService.isCreateByLogin(courseInfoId)){
+            throw new Exception("无权限创建和更改此班级.");
+        }
+        courseGroup.setDataTime(ZonedDateTime.now());
+        CourseInfo courseInfo=courseInfoRepository.findById(courseInfoId).get();
+        courseGroup.setCourse(courseInfo);
         return courseGroupRepository.save(courseGroup);
     }
 
     /**
      * Get all the courseGroups.
      *
-     * @param pageable the pagination information.
+     *
      * @return the list of entities.
      */
     @Override
     @Transactional(readOnly = true)
-    public Page<CourseGroup> findAll(Pageable pageable) {
+    public List<CourseGroup> findAll() throws Exception {
         log.debug("Request to get all CourseGroups");
-        return courseGroupRepository.findAll(pageable);
+        //检查用户是否创建过课程，若没有，则无查询班级权限
+        if(!courseInfoService.checkLoginName()){
+            throw new Exception("无权限查询班级.");
+        }
+        //检查用户是否创建过班级
+        String loginName = SecurityUtils.getCurrentUserLogin().get();
+        List<CourseGroup> courseGroups=courseGroupRepository.findByCourseCourseUser(loginName);
+        if(courseGroups==null || courseGroups.size()==0){
+            throw new Exception("未创建班级.");
+        }
+        return courseGroups;
     }
 
+    /**
+     * 通过课程ID检索开设的班级
+     * @param id
+     * @return
+     */
+    @Override
+    public List<CourseGroup> findByCourseId(Long id) throws Exception {
+        log.debug("Request to findByCourseName  : {}", id);
+        //检查课程是否由该用户创建，若不是，则无权限查询此课程下的班级
+        if(!courseInfoService.isCreateByLogin(id)){
+            throw new Exception("无权限查询此课程下的班级");
+        }
+
+        //检查用户是否在该课程下创建过班级
+        List<CourseGroup> courseGroups=courseGroupRepository.findByCourseId(id);
+        if(courseGroups==null && courseGroups.size()==0){
+            throw new Exception("此课程下未创建班级");
+        }
+        return courseGroups;
+    }
 
     /**
      * Get one courseGroup by id.
@@ -76,77 +124,37 @@ public class CourseGroupServiceImpl implements CourseGroupService {
      */
     @Override
     @Transactional(readOnly = true)
-    public Optional<CourseGroup> findOne(Long id) {
+    public Optional<CourseGroup> findOne(Long id) throws Exception {
         log.debug("Request to get CourseGroup : {}", id);
+        //检查班级是否存在
+        CourseGroup courseGroup=isExistCourseInfo(id);
+        //检查班级是否由该用户创建
+        String loginName = SecurityUtils.getCurrentUserLogin().get();
+        if(!courseGroup.getCourse().getCourseUser().equals(loginName)){
+            throw new Exception("无权限查询此班级.");
+        }
         return courseGroupRepository.findById(id);
     }
-
-    /**
-     * list转page
-     * @param list
-     * @param pageable
-     * @param <T>
-     * @return
-     */
-    public <T> Page<T> listConvertToPage(List<T> list, Pageable pageable) {
-        int start = (int)pageable.getOffset();
-        int end = (start + pageable.getPageSize()) > list.size() ? list.size() : ( start + pageable.getPageSize());
-        return new PageImpl<T>(list.subList(start, end), pageable, list.size());
-    }
-    /**
-     * Get one courseGroup by course.
-     *
-     * @param course the course of the entity.
-     * @return the entity.
-     *//*
-    @Override
-    @Transactional(readOnly = true)
-    public List<CourseGroup> findGroupByCourse(CourseInfo course) {
-        log.debug("Request to get CourseGroup : {}", course);
-        return courseGroupRepository.findAllByCourse(course);
-    }*/
-
-    /**
-     * 通过课程名检索课程
-     * @param course
-     * @return
-     */
-    @Override
-    public Page<CourseGroupDTO> findByCourse(String course,Pageable pageable) {
-        log.debug("Request to findByCourseName  : {}", course);
-        String sql="SELECT a.* FROM course_group a,course_info b WHERE b.course_name = '"+
-                    course+"' AND a.course_id = b.id";
-
-        List<Map<String,Object>> sqlResult=this.jdbcTemplate.queryForList(sql);
-        List<CourseGroupDTO> result = new ArrayList<CourseGroupDTO>();
-
-        if(sqlResult!=null && sqlResult.size()>0){
-            for (Map<String, Object> sqlItem : sqlResult) {
-                CourseGroupDTO item=new CourseGroupDTO();
-                item.setId(TypeUtils.castToLong(sqlItem.get("id")));
-                item.setGroupCode(TypeUtils.castToString(sqlItem.get("group_code")));
-                item.setGroupName(TypeUtils.castToString(sqlItem.get("group_name")));
-                item.setGroupCount(TypeUtils.castToInt(sqlItem.get("group_count")));
-
-                String time=TypeUtils.castToString(sqlItem.get("data_time"));
-                time=time.replace(".0",".000Z").replace(" ","T");
-                item.setDataTime(ZonedDateTime.parse(time, DateTimeFormatter.ISO_OFFSET_DATE_TIME.withZone(ZoneId.systemDefault())));
-
-                result.add(item);
-            }
-            return listConvertToPage(result,  pageable);
-        }
-        return null;
-    }
-
     /**
      * Delete the courseGroup by id.
      *
      * @param id the id of the entity.
      */
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws Exception {
         log.debug("Request to delete CourseGroup : {}", id);
+        //检查班级是否存在
+        CourseGroup courseGroup=isExistCourseInfo(id);
+        //检查班级是否由该用户创建
+        String loginName = SecurityUtils.getCurrentUserLogin().get();
+        if(!courseGroup.getCourse().getCourseUser().equals(loginName)){
+            throw new Exception("无权限删除此班级.");
+        }
+        //检查将要删除的班级是否有学生，若有则不能解散班级
+        List<StudentCourseGroup> studentCourseGroups= studentCourseGroupRepository.findByGroupId(id);
+        if(studentCourseGroups!=null &&studentCourseGroups.size()>0){
+            throw new Exception("班级存在学生信息.");
+        }
         courseGroupRepository.deleteById(id);
     }
 }
